@@ -1,6 +1,5 @@
-namespace JoyMoe.Common.Abstractions
+namespace JoyMoe.Common.Abstractions;
+
+public interface IDataEntity : IIdentifier, ITimestamp
 {
-    public interface IDataEntity : IIdentifier, ITimestamp
-    {
-    }
 }
