@@ -1,11 +1,10 @@
 using System;
 
-namespace JoyMoe.Common.Abstractions
-{
-    public interface ITimestamp
-    {
-        DateTime? CreationDate { get; set; }
+namespace JoyMoe.Common.Abstractions;
 
-        DateTime? ModificationDate { get; set; }
-    }
+public interface ITimestamp
+{
+    DateTime? CreationDate { get; set; }
+
+    DateTime? ModificationDate { get; set; }
 }
