@@ -3,46 +3,50 @@ using Microsoft.AspNetCore.Mvc.Abstractions;
 using Microsoft.AspNetCore.Mvc.ActionConstraints;
 using Microsoft.AspNetCore.Routing;
 
-namespace JoyMoe.Common.Attributes
+namespace JoyMoe.Common.Attributes;
+
+/// <summary>
+/// Represents an attribute that specifies which Form Data Field an action method will respond to.
+/// </summary>
+public sealed class FormValueRequiredAttribute : ActionMethodSelectorAttribute
 {
-    /// <summary>
-    /// Represents an attribute that specifies which Form Data Field an action method will respond to.
-    /// </summary>
-    public sealed class FormValueRequiredAttribute : ActionMethodSelectorAttribute
+    private readonly string _name;
+
+    public FormValueRequiredAttribute(string name)
     {
-        private readonly string _name;
+        _name = name;
+    }
 
-        public FormValueRequiredAttribute(string name)
+    public override bool IsValidForRequest(RouteContext routeContext, ActionDescriptor action)
+    {
+        if (routeContext == null)
         {
-            _name = name;
+            throw new ArgumentNullException(nameof(routeContext));
         }
 
-        public override bool IsValidForRequest(RouteContext routeContext, ActionDescriptor action)
+        if (string.Equals(routeContext.HttpContext.Request.Method, "GET",
+                          StringComparison.InvariantCultureIgnoreCase) ||
+            string.Equals(routeContext.HttpContext.Request.Method, "HEAD",
+                          StringComparison.InvariantCultureIgnoreCase) ||
+            string.Equals(routeContext.HttpContext.Request.Method, "DELETE",
+                          StringComparison.InvariantCultureIgnoreCase) ||
+            string.Equals(routeContext.HttpContext.Request.Method, "TRACE",
+                          StringComparison.InvariantCultureIgnoreCase))
         {
-            if (routeContext == null)
-            {
-                throw new ArgumentNullException(nameof(routeContext));
-            }
-
-            if (string.Equals(routeContext.HttpContext.Request.Method, "GET", StringComparison.InvariantCultureIgnoreCase) ||
-                string.Equals(routeContext.HttpContext.Request.Method, "HEAD", StringComparison.InvariantCultureIgnoreCase) ||
-                string.Equals(routeContext.HttpContext.Request.Method, "DELETE", StringComparison.InvariantCultureIgnoreCase) ||
-                string.Equals(routeContext.HttpContext.Request.Method, "TRACE", StringComparison.InvariantCultureIgnoreCase))
-            {
-                return false;
-            }
-
-            if (string.IsNullOrEmpty(routeContext.HttpContext.Request.ContentType))
-            {
-                return false;
-            }
-
-            if (!routeContext.HttpContext.Request.ContentType.StartsWith("application/x-www-form-urlencoded", StringComparison.InvariantCultureIgnoreCase))
-            {
-                return false;
-            }
-
-            return !string.IsNullOrEmpty(routeContext.HttpContext.Request.Form[_name]);
+            return false;
         }
+
+        if (string.IsNullOrEmpty(routeContext.HttpContext.Request.ContentType))
+        {
+            return false;
+        }
+
+        if (!routeContext.HttpContext.Request.ContentType.StartsWith("application/x-www-form-urlencoded",
+                                                                     StringComparison.InvariantCultureIgnoreCase))
+        {
+            return false;
+        }
+
+        return !string.IsNullOrEmpty(routeContext.HttpContext.Request.Form[_name]);
     }
 }
