@@ -2,29 +2,28 @@ using System;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
 
-namespace JoyMoe.Common.Attributes
+namespace JoyMoe.Common.Attributes;
+
+/// <summary>
+/// Provides URL validation for an array.
+/// </summary>
+[AttributeUsage(AttributeTargets.Property | AttributeTargets.Field | AttributeTargets.Parameter)]
+public sealed class UrlsAttribute : DataTypeAttribute
 {
-    /// <summary>
-    /// Provides URL validation for an array.
-    /// </summary>
-    [AttributeUsage(AttributeTargets.Property | AttributeTargets.Field | AttributeTargets.Parameter)]
-    public sealed class UrlsAttribute : DataTypeAttribute
+    private readonly UrlAttribute _url = new();
+
+    public UrlsAttribute() : base(DataType.Url)
     {
-        private readonly UrlAttribute _url = new();
+        ErrorMessage = "The {0} field is not a valid fully-qualified http, https, or ftp URL.";
+    }
 
-        public UrlsAttribute() : base(DataType.Url)
+    public override bool IsValid(object? value)
+    {
+        if (!(value is string[] valueAsStringArray))
         {
-            ErrorMessage = "The {0} field is not a valid fully-qualified http, https, or ftp URL.";
+            return true;
         }
 
-        public override bool IsValid(object? value)
-        {
-            if (!(value is string[] valueAsStringArray))
-            {
-                return true;
-            }
-
-            return valueAsStringArray.All(valueAsString => _url.IsValid(valueAsString));
-        }
+        return valueAsStringArray.All(valueAsString => _url.IsValid(valueAsString));
     }
 }
