@@ -1,11 +1,11 @@
 // ReSharper disable once CheckNamespace
-namespace System.Reflection
+
+namespace System.Reflection;
+
+internal static class CustomAttributeExtensions
 {
-    internal static class CustomAttributeExtensions
+    public static bool HasCustomAttribute<T>(this MemberInfo element, bool inherit) where T : Attribute
     {
-        public static bool HasCustomAttribute<T>(this MemberInfo element, bool inherit) where T : Attribute
-        {
-            return Attribute.IsDefined(element, typeof(T), inherit);
-        }
+        return Attribute.IsDefined(element, typeof(T), inherit);
     }
 }
