@@ -1,13 +1,12 @@
-namespace JoyMoe.Common.Data
-{
-    public static class RepositoryExtensions
-    {
-        public static IRepository<TEntity> IgnoreQueryFilters<TEntity>(this IRepository<TEntity> repository)
-            where TEntity : class
-        {
-            repository.IgnoreQueryFilters = true;
+namespace JoyMoe.Common.Data;
 
-            return repository;
-        }
+public static class RepositoryExtensions
+{
+    public static IRepository<TEntity> IgnoreQueryFilters<TEntity>(this IRepository<TEntity> repository)
+        where TEntity : class
+    {
+        repository.IgnoreQueryFilters = true;
+
+        return repository;
     }
 }
