@@ -6,32 +6,31 @@ using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Options;
 
 // ReSharper disable once CheckNamespace
-namespace Microsoft.Extensions.DependencyInjection
+namespace Microsoft.Extensions.DependencyInjection;
+
+public static class DiagnosticsMvcBuilderExtensions
 {
-    public static class DiagnosticsMvcBuilderExtensions
+    public static IServiceCollection AddApiProblemDetailsFactory(this IServiceCollection services)
     {
-        public static IServiceCollection AddApiProblemDetailsFactory(this IServiceCollection services)
+        if (services == null)
         {
-            if (services == null)
-            {
-                throw new ArgumentNullException(nameof(services));
-            }
-
-            services.TryAddSingleton<ProblemDetailsFactory, ApiProblemDetailsFactory>();
-
-            return services;
+            throw new ArgumentNullException(nameof(services));
         }
 
-        public static IHealthChecksBuilder AddRfcHealthChecks(this IServiceCollection services)
+        services.TryAddSingleton<ProblemDetailsFactory, ApiProblemDetailsFactory>();
+
+        return services;
+    }
+
+    public static IHealthChecksBuilder AddRfcHealthChecks(this IServiceCollection services)
+    {
+        if (services == null)
         {
-            if (services == null)
-            {
-                throw new ArgumentNullException(nameof(services));
-            }
-
-            services.TryAddSingleton<IPostConfigureOptions<HealthCheckOptions>, RfcHealthCheckWriterPostConfigureOptions>();
-
-            return services.AddHealthChecks();
+            throw new ArgumentNullException(nameof(services));
         }
+
+        services.TryAddSingleton<IPostConfigureOptions<HealthCheckOptions>, RfcHealthCheckWriterPostConfigureOptions>();
+
+        return services.AddHealthChecks();
     }
 }
