@@ -1,19 +1,18 @@
 using System;
 
 // ReSharper disable once CheckNamespace
-namespace Microsoft.Extensions.Diagnostics.HealthChecks
+namespace Microsoft.Extensions.Diagnostics.HealthChecks;
+
+public static class RfcHealthCheckExtensions
 {
-    public static class RfcHealthCheckExtensions
+    public static string ToRfcStatusString(this HealthStatus status)
     {
-        public static string ToRfcStatusString(this HealthStatus status)
+        return status switch
         {
-            return status switch
-            {
-                HealthStatus.Unhealthy => "fail",
-                HealthStatus.Degraded => "warn",
-                HealthStatus.Healthy => "pass",
-                _ => throw new ArgumentOutOfRangeException(nameof(status), status, null)
-            };
-        }
+            HealthStatus.Unhealthy => "fail",
+            HealthStatus.Degraded  => "warn",
+            HealthStatus.Healthy   => "pass",
+            _                      => throw new ArgumentOutOfRangeException(nameof(status), status, null)
+        };
     }
 }
