@@ -1,13 +1,12 @@
 using Microsoft.AspNetCore.Diagnostics.HealthChecks;
 using Microsoft.Extensions.Options;
 
-namespace JoyMoe.Common.Diagnostics
+namespace JoyMoe.Common.Diagnostics;
+
+public class RfcHealthCheckWriterPostConfigureOptions : IPostConfigureOptions<HealthCheckOptions>
 {
-    public class RfcHealthCheckWriterPostConfigureOptions : IPostConfigureOptions<HealthCheckOptions>
+    public void PostConfigure(string name, HealthCheckOptions options)
     {
-        public void PostConfigure(string name, HealthCheckOptions options)
-        {
-            options.ResponseWriter = RfcHealthCheckWriter.WriteResponse;
-        }
+        options.ResponseWriter = RfcHealthCheckWriter.WriteResponse;
     }
 }
