@@ -7,260 +7,266 @@ using JoyMoe.Common.Abstractions;
 using JoyMoe.Common.Data;
 using Microsoft.AspNetCore.Mvc;
 
-namespace JoyMoe.Common.Mvc.Api
+namespace JoyMoe.Common.Mvc.Api;
+
+[ApiController]
+[GenericController]
+[Route("api/[controller]")]
+public class GenericController<TEntity, TRequest, TResponse> : ControllerBase
+    where TEntity : class, IDataEntity
+    where TRequest : class, IIdentifier
+    where TResponse : class, IIdentifier
 {
-    [ApiController]
-    [GenericController]
-    [Route("api/[controller]")]
-    public class GenericController<TEntity, TRequest, TResponse> : ControllerBase
-        where TEntity : class, IDataEntity
-        where TRequest : class, IIdentifier
-        where TResponse : class, IIdentifier
+    private readonly IRepository<TEntity>                   _repository;
+    private readonly IGenericControllerInterceptor<TEntity> _interceptor;
+    private readonly IMapper                                _mapper;
+
+    public GenericController(IRepository<TEntity> repository, IGenericControllerInterceptor<TEntity> interceptor,
+                             IMapper              mapper)
     {
-        private readonly IRepository<TEntity> _repository;
-        private readonly IGenericControllerInterceptor<TEntity> _interceptor;
-        private readonly IMapper _mapper;
+        _repository  = repository;
+        _interceptor = interceptor;
+        _mapper      = mapper;
+    }
 
-        public GenericController(IRepository<TEntity> repository, IGenericControllerInterceptor<TEntity> interceptor, IMapper mapper)
+    [HttpGet]
+    public async Task<ActionResult<PaginationResponse<long, TResponse>>> Query(
+        [FromQuery] long? cursor, [FromQuery] int size = 10)
+    {
+        if (!ModelState.IsValid)
         {
-            _repository = repository;
-            _interceptor = interceptor;
-            _mapper = mapper;
+            return UnprocessableEntity(ModelState);
         }
 
-        [HttpGet]
-        public async Task<ActionResult<PaginationResponse<long, TResponse>>> Query([FromQuery] long? cursor, [FromQuery] int size = 10)
+        var result = await _interceptor.Query(HttpContext, User, predicate => _query(cursor, size, predicate))
+                                       .ConfigureAwait(false);
+
+        return _mapResponse(result);
+    }
+
+    private async Task<ActionResult<PaginationResponse<long, TEntity>>> _query(
+        long? cursor, int size, Expression<Func<TEntity, bool>>? predicate)
+    {
+        return await _repository
+                    .PaginateAsync(e => e.Id, predicate, cursor, size)
+                    .ConfigureAwait(false);
+    }
+
+    [HttpGet("{id}")]
+    public async Task<ActionResult<TResponse>> Find(long id)
+    {
+        if (!ModelState.IsValid)
         {
-            if (!ModelState.IsValid)
-            {
-                return UnprocessableEntity(ModelState);
-            }
-
-            var result = await _interceptor.Query(HttpContext, User, predicate => _query(cursor, size, predicate)).ConfigureAwait(false);
-
-            return _mapResponse(result);
+            return UnprocessableEntity(ModelState);
         }
 
-        private async Task<ActionResult<PaginationResponse<long, TEntity>>> _query(long? cursor, int size, Expression<Func<TEntity, bool>>? predicate)
+        var result = await _interceptor.Find(HttpContext, User, () => _find(id)).ConfigureAwait(false);
+
+        return _mapResponse(result);
+    }
+
+    private async Task<ActionResult<TEntity>> _find(long id)
+    {
+        var entity = await _repository.FindAsync(e => e.Id, id).ConfigureAwait(false);
+
+        if (entity == null)
         {
-            return await _repository
-                .PaginateAsync(e => e.Id,predicate, cursor, size)
-                .ConfigureAwait(false);
+            return NotFound();
         }
 
-        [HttpGet("{id}")]
-        public async Task<ActionResult<TResponse>> Find(long id)
+        return entity;
+    }
+
+    [HttpPost]
+    public async Task<ActionResult<TResponse>> Create([FromBody] TRequest? request)
+    {
+        if (request == null)
         {
-            if (!ModelState.IsValid)
-            {
-                return UnprocessableEntity(ModelState);
-            }
-
-            var result = await _interceptor.Find(HttpContext, User, () => _find(id)).ConfigureAwait(false);
-
-            return _mapResponse(result);
+            return BadRequest();
         }
 
-        private async Task<ActionResult<TEntity>> _find(long id)
+        if (!ModelState.IsValid)
         {
-            var entity = await _repository.FindAsync(e => e.Id, id).ConfigureAwait(false);
+            return UnprocessableEntity(ModelState);
+        }
 
-            if (entity == null)
-            {
-                return NotFound();
-            }
+        request.Id = default;
 
+        var entity = _mapRequest(request);
+
+        var result = await _interceptor.Create(HttpContext, User, entity, _create).ConfigureAwait(false);
+
+        return _mapResponse(result);
+    }
+
+    private async Task<ActionResult<TEntity>> _create(TEntity entity)
+    {
+        await _repository.AddAsync(entity).ConfigureAwait(false);
+
+        if (await _repository.CommitAsync().ConfigureAwait(false) == 0)
+        {
+            return Problem();
+        }
+
+        return CreatedAtAction("Find", new { id = entity.Id }, entity);
+    }
+
+    [HttpPut("{id}")]
+    public async Task<ActionResult<TResponse>> Update(long id, [FromBody] TRequest? request)
+    {
+        if (request == null)
+        {
+            return BadRequest();
+        }
+
+        if (!ModelState.IsValid)
+        {
+            return UnprocessableEntity(ModelState);
+        }
+
+        if (id != request.Id)
+        {
+            return BadRequest();
+        }
+
+        var entity = _mapRequest(request);
+
+        var result = await _interceptor.Update(HttpContext, User, entity, _update).ConfigureAwait(false);
+
+        return _mapResponse(result);
+    }
+
+    private async Task<ActionResult<TEntity>> _update(TEntity entity)
+    {
+        await _repository.UpdateAsync(entity).ConfigureAwait(false);
+
+        if (await _repository.CommitAsync().ConfigureAwait(false) == 0)
+        {
+            return Problem();
+        }
+
+        return entity;
+    }
+
+    [HttpDelete("{id}")]
+    public async Task<ActionResult> Remove(long id)
+    {
+        if (!ModelState.IsValid)
+        {
+            return UnprocessableEntity(ModelState);
+        }
+
+        var entity = await _repository.FindAsync(e => e.Id, id).ConfigureAwait(false);
+
+        if (entity == null)
+        {
+            return NotFound();
+        }
+
+        return await _interceptor.Remove(HttpContext, User, entity, _remove).ConfigureAwait(false);
+    }
+
+    private async Task<ActionResult> _remove(TEntity entity)
+    {
+        await _repository.RemoveAsync(entity).ConfigureAwait(false);
+
+        if (await _repository.CommitAsync().ConfigureAwait(false) == 0)
+        {
+            return Problem();
+        }
+
+        return NoContent();
+    }
+
+    private TEntity _mapRequest(TRequest request)
+    {
+        if (typeof(TRequest) != typeof(TEntity)) return _mapper.Map<TEntity>(request);
+
+        if (request is TEntity entity)
+        {
             return entity;
         }
 
-        [HttpPost]
-        public async Task<ActionResult<TResponse>> Create([FromBody] TRequest? request)
+        throw new NotSupportedException();
+    }
+
+    private TResponse _mapResponse(TEntity entity)
+    {
+        if (typeof(TEntity) != typeof(TResponse)) return _mapper.Map<TResponse>(entity);
+
+        if (entity is TResponse response)
         {
-            if (request == null)
-            {
-                return BadRequest();
-            }
-
-            if (!ModelState.IsValid)
-            {
-                return UnprocessableEntity(ModelState);
-            }
-
-            request.Id = default;
-
-            var entity = _mapRequest(request);
-
-            var result = await _interceptor.Create(HttpContext, User, entity, _create).ConfigureAwait(false);
-
-            return _mapResponse(result);
+            return response;
         }
 
-        private async Task<ActionResult<TEntity>> _create(TEntity entity)
+        throw new NotSupportedException();
+    }
+
+    private ICollection<TResponse> _mapResponse(ICollection<TEntity> entities)
+    {
+        if (typeof(TEntity) != typeof(TResponse)) return _mapper.Map<ICollection<TResponse>>(entities);
+
+        if (entities is ICollection<TResponse> responses)
         {
-            await _repository.AddAsync(entity).ConfigureAwait(false);
-
-            if (await _repository.CommitAsync().ConfigureAwait(false) == 0)
-            {
-                return Problem();
-            }
-
-            return CreatedAtAction("Find", new { id = entity.Id }, entity);
+            return responses;
         }
 
-        [HttpPut("{id}")]
-        public async Task<ActionResult<TResponse>> Update(long id, [FromBody] TRequest? request)
+        throw new NotSupportedException();
+    }
+
+    private PaginationResponse<long, TResponse> _mapResponse(PaginationResponse<long, TEntity> result)
+    {
+        if (typeof(TEntity) != typeof(TResponse))
         {
-            if (request == null)
+            ICollection<TResponse>? data = null;
+
+            if (result.Data != null)
             {
-                return BadRequest();
+                data = _mapResponse(result.Data);
             }
 
-            if (!ModelState.IsValid)
+            return new PaginationResponse<long, TResponse>
             {
-                return UnprocessableEntity(ModelState);
-            }
-
-            if (id != request.Id)
-            {
-                return BadRequest();
-            }
-
-            var entity = _mapRequest(request);
-
-            var result = await _interceptor.Update(HttpContext, User, entity, _update).ConfigureAwait(false);
-
-            return _mapResponse(result);
+                Prev = result.Prev,
+                Next = result.Next,
+                Data = data
+            };
         }
 
-        private async Task<ActionResult<TEntity>> _update(TEntity entity)
+        if (result is PaginationResponse<long, TResponse> response)
         {
-            await _repository.UpdateAsync(entity).ConfigureAwait(false);
-
-            if (await _repository.CommitAsync().ConfigureAwait(false) == 0)
-            {
-                return Problem();
-            }
-
-            return entity;
+            return response;
         }
 
-        [HttpDelete("{id}")]
-        public async Task<ActionResult> Remove(long id)
+        throw new NotSupportedException();
+    }
+
+    private ActionResult<TResponse> _mapResponse(ActionResult<TEntity> result)
+    {
+        if (result.Result == null) return new ActionResult<TResponse>(_mapResponse(result.Value!));
+
+        if (result.Result is not ObjectResult { Value: TEntity entity } or) return result.Result;
+
+        or.Value = _mapResponse(entity);
+
+        return new ActionResult<TResponse>(or);
+    }
+
+    private ActionResult<PaginationResponse<long, TResponse>> _mapResponse(
+        ActionResult<PaginationResponse<long, TEntity>> result)
+    {
+        if (result.Result == null)
         {
-            if (!ModelState.IsValid)
-            {
-                return UnprocessableEntity(ModelState);
-            }
-
-            var entity = await _repository.FindAsync(e => e.Id, id).ConfigureAwait(false);
-
-            if (entity == null)
-            {
-                return NotFound();
-            }
-
-            return await _interceptor.Remove(HttpContext, User, entity, _remove).ConfigureAwait(false);
+            return new ActionResult<PaginationResponse<long, TResponse>>(_mapResponse(result.Value!));
         }
 
-        private async Task<ActionResult> _remove(TEntity entity)
+        if (result.Result is not ObjectResult { Value: PaginationResponse<long, TEntity> entities } or)
         {
-            await _repository.RemoveAsync(entity).ConfigureAwait(false);
-
-            if (await _repository.CommitAsync().ConfigureAwait(false) == 0)
-            {
-                return Problem();
-            }
-
-            return NoContent();
+            return result.Result;
         }
 
-        private TEntity _mapRequest(TRequest request)
-        {
-            if (typeof(TRequest) != typeof(TEntity)) return _mapper.Map<TEntity>(request);
+        or.Value = _mapResponse(entities);
 
-            if (request is TEntity entity)
-            {
-                return entity;
-            }
-
-            throw new NotSupportedException();
-        }
-
-        private TResponse _mapResponse(TEntity entity)
-        {
-            if (typeof(TEntity) != typeof(TResponse)) return _mapper.Map<TResponse>(entity);
-
-            if (entity is TResponse response)
-            {
-                return response;
-            }
-
-            throw new NotSupportedException();
-        }
-
-        private ICollection<TResponse> _mapResponse(ICollection<TEntity> entities)
-        {
-            if (typeof(TEntity) != typeof(TResponse)) return _mapper.Map<ICollection<TResponse>>(entities);
-
-            if (entities is ICollection<TResponse> responses)
-            {
-                return responses;
-            }
-
-            throw new NotSupportedException();
-        }
-
-        private PaginationResponse<long, TResponse> _mapResponse(PaginationResponse<long, TEntity> result)
-        {
-            if (typeof(TEntity) != typeof(TResponse))
-            {
-                ICollection<TResponse>? data = null;
-
-                if (result.Data != null)
-                {
-                    data = _mapResponse(result.Data);
-                }
-
-                return new PaginationResponse<long, TResponse>
-                {
-                    Prev = result.Prev,
-                    Next = result.Next,
-                    Data = data
-                };
-            }
-
-            if (result is PaginationResponse<long, TResponse> response)
-            {
-                return response;
-            }
-
-            throw new NotSupportedException();
-        }
-
-        private ActionResult<TResponse> _mapResponse(ActionResult<TEntity> result)
-        {
-            if (result.Result == null) return new ActionResult<TResponse>(_mapResponse(result.Value));
-
-            if (!(result.Result is ObjectResult or) || !(or.Value is TEntity entity)) return result.Result;
-
-            or.Value = _mapResponse(entity);
-
-            return new ActionResult<TResponse>(or);
-
-        }
-
-        private ActionResult<PaginationResponse<long, TResponse>> _mapResponse(ActionResult<PaginationResponse<long, TEntity>> result)
-        {
-            if (result.Result == null)
-            {
-                return new ActionResult<PaginationResponse<long, TResponse>>(_mapResponse(result.Value));
-            }
-
-            if (!(result.Result is ObjectResult or) || !(or.Value is PaginationResponse<long, TEntity> entities)) return result.Result;
-
-            or.Value = _mapResponse(entities);
-
-            return new ActionResult<PaginationResponse<long, TResponse>>(or);
-        }
+        return new ActionResult<PaginationResponse<long, TResponse>>(or);
     }
 }
