@@ -4,32 +4,31 @@ using JoyMoe.Common.Mvc.Api;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 
 // ReSharper disable once CheckNamespace
-namespace Microsoft.Extensions.DependencyInjection
+namespace Microsoft.Extensions.DependencyInjection;
+
+public static class GenericControllerMvcBuilderExtensions
 {
-    public static class GenericControllerMvcBuilderExtensions
+    public static IMvcBuilder AddGenericControllers(this IMvcBuilder mvc, Action<GenericControllerBuilder> configure)
     {
-        public static IMvcBuilder AddGenericControllers(this IMvcBuilder mvc, Action<GenericControllerBuilder> configure)
+        if (mvc == null)
         {
-            if (mvc == null)
-            {
-                throw new ArgumentNullException(nameof(mvc));
-            }
-
-            if (configure == null)
-            {
-                throw new ArgumentNullException(nameof(configure));
-            }
-
-            mvc.Services.TryAddScoped(typeof(IGenericControllerInterceptor<>), typeof(GenericControllerInterceptor<>));
-
-            var builder = new GenericControllerBuilder(mvc);
-            configure(builder);
-
-            var config = new MapperConfiguration(builder.Mapper);
-            var mapper = config.CreateMapper();
-            mvc.Services.TryAddSingleton<IMapper>(mapper);
-
-            return mvc;
+            throw new ArgumentNullException(nameof(mvc));
         }
+
+        if (configure == null)
+        {
+            throw new ArgumentNullException(nameof(configure));
+        }
+
+        mvc.Services.TryAddScoped(typeof(IGenericControllerInterceptor<>), typeof(GenericControllerInterceptor<>));
+
+        var builder = new GenericControllerBuilder(mvc);
+        configure(builder);
+
+        var config = new MapperConfiguration(builder.Mapper);
+        var mapper = config.CreateMapper();
+        mvc.Services.TryAddSingleton<IMapper>(mapper);
+
+        return mvc;
     }
 }
