@@ -6,24 +6,25 @@ using JoyMoe.Common.Abstractions;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
-namespace JoyMoe.Common.Mvc.Api
+namespace JoyMoe.Common.Mvc.Api;
+
+public interface IGenericControllerInterceptor<TEntity>
+    where TEntity : class, IDataEntity
 {
-    public interface IGenericControllerInterceptor<TEntity>
-        where TEntity : class, IDataEntity
-    {
-        Task<ActionResult<PaginationResponse<long, TEntity>>> Query(HttpContext context, ClaimsPrincipal user,
-            Func<Expression<Func<TEntity, bool>>?, Task<ActionResult<PaginationResponse<long, TEntity>>>> query);
+    Task<ActionResult<PaginationResponse<long, TEntity>>> Query(HttpContext context, ClaimsPrincipal user,
+                                                                Func<Expression<Func<TEntity, bool>>?,
+                                                                    Task<ActionResult<
+                                                                        PaginationResponse<long, TEntity>>>> query);
 
-        Task<ActionResult<TEntity>> Find(HttpContext context, ClaimsPrincipal user,
-            Func<Task<ActionResult<TEntity>>> find);
+    Task<ActionResult<TEntity>> Find(HttpContext                       context, ClaimsPrincipal user,
+                                     Func<Task<ActionResult<TEntity>>> find);
 
-        Task<ActionResult<TEntity>> Create(HttpContext context, ClaimsPrincipal user, TEntity entity,
-            Func<TEntity, Task<ActionResult<TEntity>>> create);
+    Task<ActionResult<TEntity>> Create(HttpContext context, ClaimsPrincipal user, TEntity entity,
+                                       Func<TEntity, Task<ActionResult<TEntity>>> create);
 
-        Task<ActionResult<TEntity>> Update(HttpContext context, ClaimsPrincipal user, TEntity entity,
-            Func<TEntity, Task<ActionResult<TEntity>>> update);
+    Task<ActionResult<TEntity>> Update(HttpContext context, ClaimsPrincipal user, TEntity entity,
+                                       Func<TEntity, Task<ActionResult<TEntity>>> update);
 
-        Task<ActionResult> Remove(HttpContext context, ClaimsPrincipal user, TEntity entity,
-            Func<TEntity, Task<ActionResult>> remove);
-    }
+    Task<ActionResult> Remove(HttpContext                       context, ClaimsPrincipal user, TEntity entity,
+                              Func<TEntity, Task<ActionResult>> remove);
 }
