@@ -5,67 +5,67 @@ using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.Extensions.Caching.Distributed;
 
 #nullable disable
-namespace JoyMoe.Common.Session
+namespace JoyMoe.Common.Session;
+
+/// <summary>
+/// This provides an storage mechanic to preserve identity information on the <see cref="IDistributedCache" /> while only sending a simple identifier key to the client.
+/// </summary>
+public class CachedTicketStore : ITicketStore
 {
-    /// <summary>
-    /// This provides an storage mechanic to preserve identity information on the <see cref="IDistributedCache" /> while only sending a simple identifier key to the client.
-    /// </summary>
-    public class CachedTicketStore : ITicketStore
+    private const    string            KeyPrefix = "AuthSessionStore-";
+    private readonly IDistributedCache _cache;
+
+    public CachedTicketStore(IDistributedCache cache)
     {
-        private const string KeyPrefix = "AuthSessionStore-";
-        private readonly IDistributedCache _cache;
+        _cache = cache;
+    }
 
-        public CachedTicketStore(IDistributedCache cache)
+    public async Task<string> StoreAsync(AuthenticationTicket ticket)
+    {
+        var guid = Guid.NewGuid();
+        var key  = KeyPrefix + guid;
+        await RenewAsync(key, ticket).ConfigureAwait(false);
+        return key;
+    }
+
+    public async Task RenewAsync(string key, AuthenticationTicket ticket)
+    {
+        if (ticket == null)
         {
-            _cache = cache;
+            throw new ArgumentNullException(nameof(ticket));
         }
 
-        public async Task<string> StoreAsync(AuthenticationTicket ticket)
+        var options    = new DistributedCacheEntryOptions();
+        var expiresUtc = ticket.Properties.ExpiresUtc;
+        if (expiresUtc.HasValue)
         {
-            var guid = Guid.NewGuid();
-            var key = KeyPrefix + guid;
-            await RenewAsync(key, ticket).ConfigureAwait(false);
-            return key;
+            options.SetAbsoluteExpiration(expiresUtc.Value);
         }
 
-        public async Task RenewAsync(string key, AuthenticationTicket ticket)
-        {
-            if (ticket == null)
-            {
-                throw new ArgumentNullException(nameof(ticket));
-            }
+        var val = SerializeToBytes(ticket);
+        await _cache.SetAsync(key, val, options).ConfigureAwait(false);
+    }
 
-            var options = new DistributedCacheEntryOptions();
-            var expiresUtc = ticket.Properties.ExpiresUtc;
-            if (expiresUtc.HasValue)
-            {
-                options.SetAbsoluteExpiration(expiresUtc.Value);
-            }
-            var val = SerializeToBytes(ticket);
-            await _cache.SetAsync(key, val, options).ConfigureAwait(false);
-        }
+    public async Task<AuthenticationTicket> RetrieveAsync(string key)
+    {
+        var bytes  = await _cache.GetAsync(key).ConfigureAwait(false);
+        var ticket = DeserializeFromBytes(bytes);
+        return ticket;
+    }
 
-        public async Task<AuthenticationTicket> RetrieveAsync(string key)
-        {
-            var bytes = await _cache.GetAsync(key).ConfigureAwait(false);
-            var ticket = DeserializeFromBytes(bytes);
-            return ticket;
-        }
+    public async Task RemoveAsync(string key)
+    {
+        await _cache.RemoveAsync(key).ConfigureAwait(false);
+    }
 
-        public async Task RemoveAsync(string key)
-        {
-            await _cache.RemoveAsync(key).ConfigureAwait(false);
-        }
+    private static byte[] SerializeToBytes(AuthenticationTicket source)
+    {
+        return TicketSerializer.Default.Serialize(source);
+    }
 
-        private static byte[] SerializeToBytes(AuthenticationTicket source)
-        {
-            return TicketSerializer.Default.Serialize(source);
-        }
-
-        private static AuthenticationTicket DeserializeFromBytes(byte[] source)
-        {
-            return source == null ? null : TicketSerializer.Default.Deserialize(source);
-        }
+    private static AuthenticationTicket DeserializeFromBytes(byte[] source)
+    {
+        return source == null ? null : TicketSerializer.Default.Deserialize(source);
     }
 }
 #nullable restore
