@@ -3,62 +3,61 @@ using System.Threading.Tasks;
 using Automatonymous;
 using JoyMoe.Common.Abstractions;
 
-namespace JoyMoe.Common.Workflow
+namespace JoyMoe.Common.Workflow;
+
+public class EventTriggerObserver<TI> : EventObserver<TI> where TI : class, IStateful
 {
-    public class EventTriggerObserver<TI> : EventObserver<TI> where TI : class, IStateful
+    public Task PreExecute(EventContext<TI> context)
     {
-        public Task PreExecute(EventContext<TI> context)
+        return Task.CompletedTask;
+    }
+
+    public Task PreExecute<T>(EventContext<TI, T> context)
+    {
+        if (context == null)
         {
-            return Task.CompletedTask;
+            throw new ArgumentNullException(nameof(context));
         }
 
-        public Task PreExecute<T>(EventContext<TI, T> context)
+        if (context.Data is IEventData data)
         {
-            if (context == null)
+            if (string.IsNullOrWhiteSpace(data.Jockey) || data.JockeyId == 0)
             {
-                throw new ArgumentNullException(nameof(context));
+                throw new EventExecutionException();
             }
 
-            if (context.Data is IEventData data)
+            context.Instance.Note            = data.Note;
+            context.Instance.LastUpdatedById = data.JockeyId;
+            context.Instance.LastUpdatedBy   = data.Jockey;
+        }
+        else
+        {
+            if (context.Event.Name != nameof(AutomatonymousStateMachine<TI>.Initial))
             {
-                if (string.IsNullOrWhiteSpace(data.Jockey) || data.JockeyId == 0)
-                {
-                    throw new EventExecutionException();
-                }
-
-                context.Instance.Note = data.Note;
-                context.Instance.LastUpdatedById = data.JockeyId;
-                context.Instance.LastUpdatedBy = data.Jockey;
+                throw new EventExecutionException();
             }
-            else
-            {
-                if (context.Event.Name != nameof(AutomatonymousStateMachine<TI>.Initial))
-                {
-                    throw new EventExecutionException();
-                }
-            }
-
-            return Task.CompletedTask;
         }
 
-        public Task PostExecute(EventContext<TI> context)
-        {
-            return Task.CompletedTask;
-        }
+        return Task.CompletedTask;
+    }
 
-        public Task PostExecute<T>(EventContext<TI, T> context)
-        {
-            return Task.CompletedTask;
-        }
+    public Task PostExecute(EventContext<TI> context)
+    {
+        return Task.CompletedTask;
+    }
 
-        public Task ExecuteFault(EventContext<TI> context, Exception exception)
-        {
-            return Task.CompletedTask;
-        }
+    public Task PostExecute<T>(EventContext<TI, T> context)
+    {
+        return Task.CompletedTask;
+    }
 
-        public Task ExecuteFault<T>(EventContext<TI, T> context, Exception exception)
-        {
-            return Task.CompletedTask;
-        }
+    public Task ExecuteFault(EventContext<TI> context, Exception exception)
+    {
+        return Task.CompletedTask;
+    }
+
+    public Task ExecuteFault<T>(EventContext<TI, T> context, Exception exception)
+    {
+        return Task.CompletedTask;
     }
 }
